feat: :tada: Initial configuration of ecommerce
Agregamos los primeros recursos que usaremos, es decir que las fuentes, además de que empezamos con la separación de las diferentes espacios que usaremos (Storage, Img, fonts)

feat: :wrench: plantilla a usar
Agregamos las diferentes vistas que tendremos, los cuales son la pagina principal, las tarjetas y el balance o dinero que tenemos, teniamos y podriamos tener.

chore: :lipstick: Importing stlyes and structure.

I add the importation of the style in the different pages, and i give the structure in Index

style: :lipstick: Desing and structure.

In index, i put the icon of inveto, and start with the desing of the main page, when i have the top, everything gonna be more easy.
In index, I add the image of the filtrer of the user, so later i gonna solve the problem that it dont fit where it supposed to be. In style i add the different margins of the page to dont have empty spaces. In variables y add the style of letter that have the tittle (The hello world is just the beginning)
I solve the problem that makes the text and the icons get copiled in a stock. Now i what i have to do it's get text of INVETO next to the icon of the left, and solve the different problems of the sizes.
So.. here i put the image of the filtrer but in color white and i put the coditions for the header (size of letter, the images and a new color for the letter).

feat: :lipstick: Design of the background and others.

I finally move the text to the left close to the icon, so the part of the header it's done. Solved.
I add the image of the background, and now i gonna fix it later.
Here i put the explain more in detail, so you can defer what is everything and what space is getting everything, you can defer between meanwhile the different colors in each one. In Index i put the nav for the introduction for the user (Hi, Arthur)
Progress. I solve the problem that makes the letters dont fit, so now i'm doing the part of savings and assets, to after do the part of activities.
I add the images of the bar, and new conditions for the format. And i forget add in the last commit that i add a new image. In this commit i add a new font for the text

feat: :lipstick: Design of the background and others.

I finally move the text to the left close to the icon, so the part of the header it's done. Solved.
I add the image of the background, and now i gonna fix it later.
Here i put the explain more in detail, so you can defer what is everything and what space is getting everything, you can defer between meanwhile the different colors in each one. In Index i put the nav for the introduction for the user (Hi, Arthur)
Progress. I solve the problem that makes the letters dont fit, so now i'm doing the part of savings and assets, to after do the part of activities.
I add the images of the bar, and new conditions for the format. And i forget add in the last commit that i add a new image. In this commit i add a new font for the text

style: :lipstick: Process of the left and main side.

I solve the problem of the bars in the left side, so i save progress. Now i have to solve the right side.
I remove the message of the correction of Chat gpt (He correct two spaces between the columns)
So, i finish the part of  above. The bars with colors are done. Now i process to make the part of main, and the footer.
Progress of the section of activities.

style: :dizzy: Main and new functions in process

I start doing the part of main (everything with relation of activities and transactions)
I add a "new" function of section for contain the first information of the main, and later for the second part i gonna use another function of section
I add a transition when you click or pass the click above of it, it get a color.

feat: :lipstick: Footer done
So i made the footer of the page, and i add some images for the other pages.

So i add space between the different parts (main and footer in the css) because i was getting confused between each and other. And about the code, i made the first transaction.

Progress, now i have the bar of Main, i have to solve error in the spaces between everything:c

Style: :lipstick: FIRST PAGE DONE AND REDIRECTION.

I FINISH THE FIRST PAGE :D
FINISHED, i correct all the extra colors of the differentiaton of the divs and everything.
And i add the redirection links to the other pages.

style: :lipstick: Main page done with customization.

Translate of a word
First page customization
I add new pictures for the list in the middle, and add new sections. And im going to change the background of the site
I change the background of the main site. And change a wrong icon.
I finished my main page, customized, changing values, colors, backgrounds and sizes.

docs: :memo: Update of the video.
Updating of the video prof of the project.

fix: :AMBULANCE: Rename functions.
Change the names of everything for solve errors.