feat: :tada: configuración inicial ecommerce
Agregamos los primeros recursos que usaremos, es decir que las fuentes, además de que empezamos con la separación de las diferentes espacios que usaremos (Storage, Img, fonts)

feat: :wrench: plantilla a usar
Agregamos las diferentes vistas que tendremos, los cuales son la pagina principal, las tarjetas y el balance o dinero que tenemos, teniamos y podriamos tener.

hash: e1e1d45

feat: :lipstick: Add the images and structure.

Well, everything now is in english. So... I add the different images that i gonna use meanwhile the development of this project.
I create the css's of the structure that i gonna use.
I add the different colors that i need.
New image

hash: 15bb93d

chore: :lipstick: Importing stlyes and structure.

I add the importation of the style in the different pages, and i give the structure in Index

hash: 95c1cec

style: :lipstick: Desing and structure.

In index, i put the icon of inveto, and start with the desing of the main page, when i have the top, everything gonna be more easy.
In index, I add the image of the filtrer of the user, so later i gonna solve the problem that it dont fit where it supposed to be. In style i add the different margins of the page to dont have empty spaces. In variables y add the style of letter that have the tittle (The hello world is just the beginning)
I solve the problem that makes the text and the icons get copiled in a stock. Now i what i have to do it's get text of INVETO next to the icon of the left, and solve the different problems of the sizes.
So.. here i put the image of the filtrer but in color white and i put the coditions for the header (size of letter, the images and a new color for the letter).

hash: 4b4c8d7

feat: :lipstick: Design of the background and others.

I finally move the text to the left close to the icon, so the part of the header it's done. Solved.
I add the image of the background, and now i gonna fix it later.
Here i put the explain more in detail, so you can defer what is everything and what space is getting everything, you can defer between meanwhile the different colors in each one. In Index i put the nav for the introduction for the user (Hi, Arthur)
Progress. I solve the problem that makes the letters dont fit, so now i'm doing the part of savings and assets, to after do the part of activities.
I add the images of the bar, and new conditions for the format. And i forget add in the last commit that i add a new image. In this commit i add a new font for the text

hash: fbfa658

style: :lipstick: Fonts and colors of the page.

I add another fonts to use, in this case i use it for the numbers in the part of balance of the main page or HOME. And i solve to put SAVINGS and ASSETS in two different columns and make it works, so now i have to solve the problem of the bars that dont make what i want :c
I put the color of the letters and the numbers.

hash: cdf24fe

style: :lipstick: Process of the left and main side.

I solve the problem of the bars in the left side, so i save progress. Now i have to solve the right side.
I remove the message of the correction of Chat gpt (He correct two spaces between the columns)
So, i finish the part of  above. The bars with colors are done. Now i process to make the part of main, and the footer.
Progress of the section of activities.

hash: b6b5e37

style: :dizzy: Main and new functions in process

I start doing the part of main (everything with relation of activities and transactions)
I add a "new" function of section for contain the first information of the main, and later for the second part i gonna use another function of section
I add a transition when you click or pass the click above of it, it get a color.

hash: f39fc1c

feat: :lipstick: Footer done
So i made the footer of the page, and i add some images for the other pages.

So i add space between the different parts (main and footer in the css) because i was getting confused between each and other. And about the code, i made the first transaction.

Progress, now i have the bar of Main, i have to solve error in the spaces between everything:c

hash: 06bfafa

Style: :lipstick: FIRST PAGE DONE AND REDIRECTION.

I FINISH THE FIRST PAGE :D
FINISHED, i correct all the extra colors of the differentiaton of the divs and everything.
And i add the redirection links to the other pages.

hash: 589f10e

style: :lipstick: Main page done with customization.

Translate of a word
First page customization
I add new pictures for the list in the middle, and add new sections. And im going to change the background of the site
I change the background of the main site. And change a wrong icon.
I finished my main page, customized, changing values, colors, backgrounds and sizes.

hash: 6b22a8c

Style :lipstick: Change colors and personalitation

Customing the Investing and balance pages.
Customitation of the Balance page and putting the section of the footer.
Change of the color of the icons footer.
change the colors and background
I add color to the transactions