style: :lipstick: Desing and structure.
In index, i put the icon of inveto, and start with the desing of the main page, when i have the top, everything gonna be more easy.
In index, I add the image of the filtrer of the user, so later i gonna solve the problem that it dont fit where it supposed to be. In style i add the different margins of the page to dont have empty spaces. In variables y add the style of letter that have the tittle (The hello world is just the beginning)
I solve the problem that makes the text and the icons get copiled in a stock. Now i what i have to do it's get text of INVETO next to the icon of the left, and solve the different problems of the sizes.
So.. here i put the image of the filtrer but in color white and i put the coditions for the header (size of letter, the images and a new color for the letter).