feat: :lipstick: Design of the background and others.
I finally move the text to the left close to the icon, so the part of the header it's done. Solved.
I add the image of the background, and now i gonna fix it later.
Here i put the explain more in detail, so you can defer what is everything and what space is getting everything, you can defer between meanwhile the different colors in each one. In Index i put the nav for the introduction for the user (Hi, Arthur)
Progress. I solve the problem that makes the letters dont fit, so now i'm doing the part of savings and assets, to after do the part of activities.
I add the images of the bar, and new conditions for the format. And i forget add in the last commit that i add a new image. In this commit i add a new font for the text