style: :lipstick: Process of the left and main side.
I solve the problem of the bars in the left side, so i save progress. Now i have to solve the right side.
I remove the message of the correction of Chat gpt (He correct two spaces between the columns)
So, i finish the part of  above. The bars with colors are done. Now i process to make the part of main, and the footer.
Progress of the section of activities.