feat: :lipstick: Footer done So i made the footer of the page, and i add some images for the other pages.
So i add space between the different parts (main and footer in the css) because i was getting confused between each and other. And about the code, i made the first transaction.

Progress, now i have the bar of Main, i have to solve error in the spaces between everything:c