Style: :lipstick: Page design and space errors solved.
Progress (Change the spaces and add a margin in main)
Progress of the main. I have to solve the problem of the height and width of the spaces of the cards.
Im working in the solving of the problem of the pig in the nav. And after solve that, im gonna work in the part of the cards.
Solving little errors in the code, that makes the logo of the cards go to the middle when it have to be in the left of the card