style: :lipstick: Main page done with customization.
Translate of a word
First page customization
I add new pictures for the list in the middle, and add new sections. And im going to change the background of the site
I change the background of the main site. And change a wrong icon.
I finished my main page, customized, changing values, colors, backgrounds and sizes.